requirements.txt
gspread
jira-python
launchpadlib